+++ packages/analytics/lib/analytics.dart
@@ -1,15 +1,19 @@
+import 'package:deriv_rudderstack/deriv_rudderstack.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_segment/flutter_segment.dart';
 
 import 'analytics_route_observer.dart';
 
-/// Class that collects and send analytical information to "Firebase" and "Segment"
+/// Class that collects and send analytical information to "Firebase" and
+/// "RudderStack".
 class Analytics {
+  /// Initialises
+  factory Analytics() => _instance;
+
   Analytics._internal();
 
   /// A public instance of the class [Analytics].
-  static final Analytics instance = Analytics._internal();
+  static final Analytics _instance = Analytics._internal();
 
   /// List contains ignored routes/screen names
   List<String> _ignoredRoutes = <String>[];
@@ -20,19 +24,18 @@ class Analytics {
   AnalyticsRouteObserver observer;
 
   /// Initialises the "Analytics".
-  /// Sets the device-token to "Segment".
+  /// Sets the device-token to "RudderStack".
   /// bool [isEnabled] enables or disables "Analytics".
-  void init({@required String deviceToken, @required bool isEnabled}) {
+  Future<void> init({@required bool isEnabled}) async {
     _firebaseAnalytics = FirebaseAnalytics();
     observer = AnalyticsRouteObserver(onNewRoute: _newRouteHandler);
 
     // Enable or disable the analytics on this device.
-    _firebaseAnalytics.setAnalyticsCollectionEnabled(isEnabled);
-    isEnabled ? Segment.enable() : Segment.disable();
+    await _firebaseAnalytics.setAnalyticsCollectionEnabled(isEnabled);
 
-    if (deviceToken != null) {
-      _setSegmentDeviceToken(deviceToken);
-    }
+    isEnabled
+        ? await DerivRudderstack().enable()
+        : await DerivRudderstack().disable();
   }
 
   /// Captures `screen_view` event on route changes.
@@ -47,17 +50,17 @@ class Analytics {
   void logAppOpened() {
     _firebaseAnalytics?.logAppOpen();
 
-    Segment.track(eventName: 'Application Opened');
+    DerivRudderstack().track(eventName: 'Application Opened');
   }
 
   /// Captures `Application Backgrounded` event when the app goes to background.
   void logAppBackgrounded() {
-    Segment.track(eventName: 'Application Backgrounded');
+    DerivRudderstack().track(eventName: 'Application Backgrounded');
   }
 
   /// Captures `Application Crashed` event when the app is crashed.
   void logAppCrashed() {
-    Segment.track(eventName: 'Application Crashed');
+    DerivRudderstack().track(eventName: 'Application Crashed');
   }
 
   /// Used to capture information about current screen in use.
@@ -70,20 +73,23 @@ class Analytics {
     }
     _firebaseAnalytics?.setCurrentScreen(screenName: screenName);
 
-    Segment.screen(
+    DerivRudderstack().screen(
       screenName: screenName,
       properties: properties,
     );
   }
 
   /// Captures `login` event upon a successful user log in.
-  void logLoginEvent(int userId) {
-    _setFirebaseUserId(userId.toString());
-    _firebaseAnalytics?.logLogin();
+  Future<void> logLoginEvent(
+      {@required String deviceToken, @required int userId}) async {
+    await _setFirebaseUserId(userId.toString());
+    await _firebaseAnalytics?.logLogin();
 
-    Segment.identify(
-      userId: userId.toString(),
-    );
+    if (deviceToken != null) {
+      await _setRudderStackDeviceToken(deviceToken);
+    }
+
+    await DerivRudderstack().identify(userId: userId.toString());
   }
 
   /// Captures `logout` event when the user logs out.
@@ -91,15 +97,23 @@ class Analytics {
     _firebaseAnalytics?.logEvent(name: 'logout');
   }
 
-  /// Sets the device-token to "Segment".
-  void _setSegmentDeviceToken(String deviceToken) =>
-      Segment.setContext(<String, dynamic>{
-        'device': <String, dynamic>{'token': deviceToken}
-      });
+  /// Sets the device-token to "RudderStack".
+  Future<void> _setRudderStackDeviceToken(String deviceToken) =>
+      DerivRudderstack().setContext(token: deviceToken);
 
   /// Sets the user id to "Firebase".
   Future<void> _setFirebaseUserId(String userId) =>
       _firebaseAnalytics?.setUserId(userId);
+
+  /// Logs push token.
+  Future<void> logPushToken(String deviceToken) async {
+    if (deviceToken != null) {
+      await _setRudderStackDeviceToken(deviceToken);
+    }
+  }
+
+  /// Should be called at logout to clear up current rudder stack data.
+  Future<void> reset() async => DerivRudderstack().reset();
 
   /// Used to log custom events to "Firebase"
   Future<void> logToFirebase({