+++ packages/deriv_technical_analysis/test/indicators/calculations/helper_indicators/high_value_indicator_test.dart
@@ -1,4 +1,4 @@
-import 'package:deriv_technical_analysis/src/indicators/calculations/helper_indicators/high_value_inidicator.dart';
+import 'package:deriv_technical_analysis/deriv_technical_analysis.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../mock_models.dart';